requirements.txt
yfinance
fastapi
pydantic
requests
uvicorn
pandas
sentence-transformers
faiss-cpu
numpy